requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
pyttsx3>=2.90
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
sentence-transformers>=2.2.0 
pypdf>=3.0.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1